added line view model file
